Make nrf52 watchdog timer be consistent with the lpc55 one
